requirements.txt
beautifulsoup4==4.11.1
Flask==2.2.1
Flask-Cors==3.0.10
GitPython==3.1.27
gunicorn==20.1.0
Markdown==3.4.1
mysql-connector==2.2.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1